Add "readwire" interface to *some* Graphene Types.
This is useful when you want to *read* the serialized
wire protocol back into objects-in-memory. Although this
is rarely used on the client side of graphene networks,
at least one particular use case (BitShares blind transfers)
exists.

To clarify: this is __bytes__ method in reverse.

The `_readwire` interface and method signature are
pretty horrid, so some redesign might be required.